+++ pkg/jsonmessage/jsonmessage.go
@@ -23,8 +23,8 @@ func (e *JSONError) Error() string {
 
 type JSONProgress struct {
 	terminalFd uintptr
-	Current    int   `json:"current,omitempty"`
-	Total      int   `json:"total,omitempty"`
+	Current    int64 `json:"current,omitempty"`
+	Total      int64 `json:"total,omitempty"`
 	Start      int64 `json:"start,omitempty"`
 }
 
@@ -64,7 +64,7 @@ func (p *JSONProgress) String() string {
 	numbersBox = fmt.Sprintf("%8v/%v", current, total)
 
 	if p.Current > 0 && p.Start > 0 && percentage < 50 {
-		fromStart := time.Now().UTC().Sub(time.Unix(int64(p.Start), 0))
+		fromStart := time.Now().UTC().Sub(time.Unix(p.Start, 0))
 		perEntry := fromStart / time.Duration(p.Current)
 		left := time.Duration(p.Total-p.Current) * perEntry
 		left = (left / time.Second) * time.Second
